Add Pathes argument as required + Convert DaysBack to negative + Add validation on deleteFiles + Fix print date before

+++ clean.ps1
@@ -1,28 +1,33 @@
 <#
     It deletes old files on a list of folder
 #>
-param($Pathes, $Daysback)
-if (-not $Pathes) {
-    $Pathes = @('C:\temp',"$HOME\OneDrive - Anheuser-Busch InBev\Desktop\temp")
-}
-if (-not $Daysback) {
-    $Daysback = "-30"
-}
 
 # Delete all Files in a specific directory that are older than X day(s)
+param (
+    [string[]]$Pathes = $(throw "Missing parameter: Pass the path list (-Pathes)"),
+    [int]$DaysBack = 60
+)
+# Put Daysback on negative form
+$DaysBack = $DaysBack - ($DaysBack * 2)
+
 function deleteFiles {
-    param($Path, $CurrentDate, $Daysback)
-    $DateToDelete = $CurrentDate.AddDays($Daysback)
+    param($Path, $DateToDelete)
+    # Avoid accidentally delete
+    if (-not($Path -Or $DateToDelete)) {
+        break
+    }
     Get-ChildItem "$Path" -Recurse | Where-Object { $_.LastWriteTime -lt $DatetoDelete } | Remove-Item
 }
 
-Write-Host("Starting deleting before $DateToDelete...")
+$CurrentDate = Get-Date
+$DateToDelete = $CurrentDate.AddDays($DaysBack)
+Write-Host("Starting deleting before", $dateToDelete.ToString("yyyy-MM-dd"))
 foreach ($Path in $Pathes) {
-    $CurrentDate = Get-Date
     if (-not(Test-Path $Path)) {
         Write-Host "(-) Path $path doesn't exist."
         continue
     }
     Write-Host("(+) Deleting files on $Path")
-    deleteFiles -Path $Path -CurrentDate $CurrentDate -Daysback $Daysback
+    deleteFiles -Path $Path -DateToDelete $DateToDelete
 }
+Write-Host("Done.")
